lc-backend/requirements.txt
supabase
Django>=4.0
celery
redis
datetime
python-decouple
stripe
dj_database_url
djangorestframework
django-cors-headers
appwrite
uuid
psycopg2-binary
schedule
gunicorn